refactor: 'storage' 모듈의 이름을 'schedule' 로 변경
