+++ feature/schedule/proguard-rules.pro
