+++ feature/schedule/src/main/AndroidManifest.xml
