+++ feature/schedule/src/main/java/com/hobbyloop/feature/schedule/StorageNavigation.kt
@@ -1,14 +1,14 @@
-package com.hobbyloop.feature.storage
+package com.hobbyloop.feature.schedule
 
 import androidx.compose.ui.graphics.Color
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 
-internal const val STORAGE_ROUTE = "storage"
+internal const val SCHEDULE_ROUTE = "storage"
 
-internal fun NavGraphBuilder.storageScreen(onContentColor: (color: Color) -> Unit) {
-    composable(route = STORAGE_ROUTE) {
-        StorageScreen(
+internal fun NavGraphBuilder.scheduleScreen(onContentColor: (color: Color) -> Unit) {
+    composable(route = SCHEDULE_ROUTE) {
+        ScheduleScreen(
             backgroundColor = Color.Blue.copy(alpha = 0.1f),
             onContentColor = onContentColor,
         )
