+++ settings.gradle.kts
@@ -28,7 +28,7 @@ include(":feature:sign-up")
 include(":feature:home")
 include(":feature:center")
 include(":feature:reservation")
-include(":feature:storage")
+include(":feature:schedule")
 include(":feature:my-page")
 include(":feature:center-detail")
 include(":feature:reservation-detail")
